% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{bic2_sim}
\alias{bic2_sim}
\title{Modified-BIC computation for a sim_mult output}
\usage{
bic2_sim(sim)
}
\arguments{
\item{sim}{A sim_mult output}
}
\value{
Return an array containing Modified-BIC  value for each simulation
}
\description{
Modified-BIC computation for a sim_mult output
}
