% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{bic_sim}
\alias{bic_sim}
\title{BIC computation for a sim_mult output}
\usage{
bic_sim(sim)
}
\arguments{
\item{sim}{A sim_mult output}
}
\value{
Return an array containing BIC value for each simulation
}
\description{
BIC computation for a sim_mult output
}
