% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{binmixmod}
\alias{binmixmod}
\title{Computes at the same time and with the same initial parameters binned CL-EM algorithm and classical EM (using Rmixmod)}
\usage{
binmixmod(data, class, R, maxit, tol, nrep)
}
\arguments{
\item{data}{Needs to a matrix of data}

\item{class}{Number of groups of the mixture models}

\item{R}{Needs to be a matrix or a numerical vector. The element (i,j) indicates the density of the j-th grid along the i-th dimension}

\item{maxit}{Maximum number of iterations EM}

\item{tol}{Tolerance of EM}

\item{nrep}{Number of different initialization of EM}
}
\value{
Binned CL-EM and classical EM algorithm output
}
\description{
Computes at the same time and with the same initial parameters binned CL-EM algorithm and classical EM (using Rmixmod)
}
