% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{binmixt_to_Rmix}
\alias{binmixt_to_Rmix}
\title{Binmixt paramter to Rmixmod parameter}
\usage{
binmixt_to_Rmix(binmixt)
}
\arguments{
\item{binmixt}{parameter
Binmixt paramter to Rmixmod parameter}
}
\description{
Binmixt paramter to Rmixmod parameter
}
