% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{build_bin_marg}
\alias{build_bin_marg}
\title{build_bin_marg}
\usage{
build_bin_marg(x, R)
}
\arguments{
\item{x}{Needs to a maxtrix of data}

\item{R}{Vector. The i-th element indicates the density of grid along the i-th dimension}
}
\description{
Build marginal binned data
}
