% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{build_bin_mext}
\alias{build_bin_mext}
\title{Build multivariate/univariate binned data using a square grid}
\usage{
build_bin_mext(x, R)
}
\arguments{
\item{x}{Needs to a matrix or a vector of data}

\item{R}{Needs to be a number. It indicates the density of grid along each dimension}

\item{sigma}{Needs to be a list of matrix}
}
\value{
The binned data and the square grid
}
\description{
Build multivariate/univariate binned data using a square grid
}
