% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{contellipse}
\alias{contellipse}
\title{Tool for drawing confidence ellipses for components of a bivariate Gaussian mixture}
\usage{
contellipse(s1, s2, mu, v, alpha, color = 1, lty, add = F)
}
\arguments{
\item{s1}{Points to draw in the first dimension}

\item{s2}{Points to draw in the second dimension}

\item{mu}{List of means}

\item{v}{List of variances}

\item{alpha}{Level of confidence}

\item{color}{Colour of the ellipses (default=1)}

\item{add}{Logical value (default=F). If TRUE, ellpses are added to the existing plot}
}
\description{
Tool for drawing confidence ellipses for components of a bivariate Gaussian mixture
}
