% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{cumul_mixt}
\alias{cumul_mixt}
\title{Univariate Gaussian mixture cdf}
\usage{
cumul_mixt(x, pi, mu, v, log = T, lower.tail = TRUE)
}
\arguments{
\item{x}{real vector}

\item{pi}{proportions}

\item{mu}{means}

\item{v}{variances}

\item{log}{Logical. If TRUE (default), it returns the logarithm}

\item{lower.tails}{Logical. If TRUE (default), it returns P[X<=x], if FALSE P[X>=x]}
}
\value{
Univariate Gaussian mixture cdf
}
\description{
Univariate Gaussian mixture cdf
}
