% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{d_mixt}
\alias{d_mixt}
\title{Density of univariate Gaussian mixture}
\usage{
d_mixt(x, pi, mu, v)
}
\arguments{
\item{x}{real vector}

\item{pi}{proportions}

\item{mu}{means}

\item{v}{variances}
}
\value{
Density of univariate Gaussian mixture
}
\description{
Density of univariate Gaussian mixture
}
