% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{d_mixt_m}
\alias{d_mixt_m}
\title{Density of multivariate Gaussian mixture}
\usage{
d_mixt_m(x, pi, mu, v)
}
\arguments{
\item{x}{real matrix}

\item{pi}{proportions}

\item{mu}{means}

\item{v}{variances}
}
\value{
Density of multivariate Gaussian mixture
}
\description{
Density of multivariate Gaussian mixture
}
