% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{em_bin_gaus_comp_nomixt}
\alias{em_bin_gaus_comp_nomixt}
\title{Binned CL-EM algorithm for multivariate binned diagonal Gaussian model}
\usage{
em_bin_gaus_comp_nomixt(bin, mu0, sigma0, tol, tim)
}
\arguments{
\item{bin}{Binned data}

\item{mu0}{Starting point for means}

\item{sigma0}{Starting point for variances}

\item{tol}{Tolerance of the algorithm}

\item{tim}{Maximum number of iterations}
}
\value{
List with original binned data, final estimates, log-likelihood and parameters sequences and medium time per iteration
}
\description{
Binned CL-EM algorithm for multivariate binned diagonal Gaussian model
}
