% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{em_dim_binned_pimu}
\alias{em_dim_binned_pimu}
\title{Inner function of hybrid binned CL-EM algorithm for proportion
and means calculation and other useful quantities}
\usage{
em_dim_binned_pimu(bin, subs, pi0, mu0, sigma0, d, raw.results)
}
\description{
Inner function of hybrid binned CL-EM algorithm for proportion
and means calculation and other useful quantities
}
