% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{em_dim_var}
\alias{em_dim_var}
\title{Inner function of hybrid binned CL-EM algorithm}
\usage{
em_dim_var(d, subs, raw.results, binned.results)
}
\description{
Inner function of hybrid binned CL-EM algorithm
}
