% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{em_raw_inner}
\alias{em_raw_inner}
\title{Inner function of hybrid binned CL-EM algorithm for raw proportion
and means calculation and other useful quantities}
\usage{
em_raw_inner(subs, pi0, mu0, sigma0)
}
\description{
Inner function of hybrid binned CL-EM algorithm for raw proportion
and means calculation and other useful quantities
}
