% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{em_subs}
\alias{em_subs}
\title{EM algorithm for univariate Gaussian mixture}
\usage{
em_subs(x, pi0, mu0, v0, tol, tim)
}
\arguments{
\item{x}{Data}

\item{pi0}{Starting point for proportions}

\item{mu0}{Starting point for means}

\item{v0}{Starting point for variances}

\item{tol}{Tolerance of the algorithm}

\item{tim}{Maximum number of iterations}
}
\value{
List with final estimates, log-likelihood sequence, BIC and medium time per iteration
}
\description{
EM algorithm for univariate Gaussian mixture
}
