% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{em_univ_hybrid}
\alias{em_univ_hybrid}
\title{Hybrid Binned-EM algorithm for univariate Gaussian mixtures}
\usage{
em_univ_hybrid(bin, subs, pi0, m0, s0, tim, tol)
}
\arguments{
\item{bin}{Univariate binned data}

\item{subs}{Subsample}

\item{pi0}{Starting point for proportions. If missing, a binned_EM algorithm for a single Gaussian is performed}

\item{tim}{Maximum number of iterations}

\item{tol}{Tolerance of the algorithm}

\item{mu0}{Starting point for means (numerical vector)}

\item{sigma0}{Starting point for variances (numerical vector)}
}
\value{
List with final estimates, log-likelihood sequence, parameters sequence, BIC and medium time per iteration
}
\description{
Hybrid Binned-EM algorithm for univariate Gaussian mixtures
}
