% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{gen_mixt}
\alias{gen_mixt}
\title{Generates a univariate gaussian mixture}
\usage{
gen_mixt(n, p, mu, v, seed)
}
\arguments{
\item{p}{Needs to a vector of probabilities}

\item{mu}{Needs to be a vector of means}

\item{v}{Needs to be a vector of variances}

\item{seed}{Fix a random seed (optional)}
}
\description{
Generates a univariate gaussian mixture
}
