% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{gen_mixt_m}
\alias{gen_mixt_m}
\title{Generates a gaussian multivariate mixture}
\usage{
gen_mixt_m(n, p, mu, sigma, seed)
}
\arguments{
\item{p}{Needs to a vector of probabilities}

\item{mu}{Needs to be a list of vector}

\item{sigma}{Needs to be a list of matrix}

\item{seed}{Fix a random seed (optional)}
}
\description{
Generates a gaussian multivariate mixture
}
