% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{logli_marg}
\alias{logli_marg}
\title{Marginal composite log-likelihood of binned diagonal Gaussian mixture}
\usage{
logli_marg(bin, pi0, mu0, sigma0)
}
\arguments{
\item{bin}{Binned data}

\item{pi0}{Proportions}

\item{mu0}{List of means}

\item{sigma0}{List of diagonal matrices}
}
\value{
Marginal composite log-likelihood for binned diagonal Gaussian mixture
}
\description{
Marginal composite log-likelihood of binned diagonal Gaussian mixture
}
