% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{logli_mixt}
\alias{logli_mixt}
\title{Univariate Gaussian mixture log-likelihood}
\usage{
logli_mixt(x, pi, mu, v)
}
\arguments{
\item{x}{real vector}

\item{pi}{proportions}

\item{mu}{means}

\item{v}{variances}
}
\value{
Univariate Gaussian mixture log-likelihood
}
\description{
Univariate Gaussian mixture log-likelihood
}
