% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{logli_mult}
\alias{logli_mult}
\title{Multinomial log-likelihood for binned univariate Gaussian mixture}
\usage{
logli_mult(bin, pi, mu, v)
}
\arguments{
\item{bin}{Binned data}

\item{pi}{proportions}

\item{mu}{means}

\item{v}{variances}
}
\value{
Multinomial log-likelihood for binned univariate Gaussian mixture
}
\description{
Multinomial log-likelihood for binned univariate Gaussian mixture
}
