% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{param_init_mclust}
\alias{param_init_mclust}
\title{Transform mclust output into a list of parameters compatible with binMixt package}
\usage{
param_init_mclust(model)
}
\arguments{
\item{model}{A mclust model}
}
\value{
List of parameters compatible with binMixt package
}
\description{
Transform mclust output into a list of parameters compatible with binMixt package
}
