% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{posterior.bin_dim2}
\alias{posterior.bin_dim2}
\title{Binned posterior probabilities  (bivariate case)}
\usage{
posterior.bin_dim2(bin, model)
}
\arguments{
\item{model}{A set of binned estimates}

\item{data}{Matrix of raw data}
}
\value{
Matrix of posterior probabilities
}
\description{
Binned posterior probabilities  (bivariate case)
}
