% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{posterior}
\alias{posterior}
\title{Raw posterior probabilities}
\usage{
posterior(data, model)
}
\arguments{
\item{data}{Matrix of raw data}

\item{model}{A set of binned estimates}
}
\value{
Matrix of posterior probabilities
}
\description{
Raw posterior probabilities
}
