% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixt.R
\name{sim_mult}
\alias{sim_mult}
\title{Build multivariate binned data using a square grid}
\usage{
sim_mult(num_sim, maxit, tol, class, nrep, R, n, pi, mu, sigma)
}
\arguments{
\item{num_sim}{Number of simulation}

\item{maxit}{Maximum number of iterations EM}

\item{tol}{Tolerance of EM}

\item{class}{Number of groups of the mixture models}

\item{nrep}{Number of different initialization of EM}

\item{R}{Needs to be a matrix or a numerical vector. The element (i,j) indicates the density of the j-th grid along the i-th dimension}

\item{n}{number of instances simulated}

\item{pi}{Proportions of the true mixture (vector)}

\item{mu}{Means of the true mixture (list of vectors)}

\item{sigma}{Variances of the true mixture (list of diagonal matrices)}
}
\value{
Binned CL-EM estimation
}
\description{
Build multivariate binned data using a square grid
}
